Move TermOrdValComparator and TermValComparator back to FieldComparator
This is done to see changes in PR easier
