+++ lucene/core/src/java/org/apache/lucene/search/SortField.java
@@ -29,8 +29,8 @@ import org.apache.lucene.search.comparators.DoubleComparator;
 import org.apache.lucene.search.comparators.FloatComparator;
 import org.apache.lucene.search.comparators.IntComparator;
 import org.apache.lucene.search.comparators.LongComparator;
-import org.apache.lucene.search.comparators.TermOrdValComparator;
-import org.apache.lucene.search.comparators.TermValComparator;
+import org.apache.lucene.search.FieldComparator.TermOrdValComparator;
+import org.apache.lucene.search.FieldComparator.TermValComparator;
 import org.apache.lucene.store.DataInput;
 import org.apache.lucene.store.DataOutput;
 import org.apache.lucene.util.BytesRef;
